+++ ts-app-demo/src/Cart.tsx
@@ -1,6 +1,6 @@
 import Spinner from "./Spinner";
 import { useNavigate } from "react-router-dom";
-import { useStore } from "./store";
+import { useStore } from "./shoeStore";
 import { CartItem, Product } from "./types/types";
 import { useEffect, useState } from "react";
